# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Jeunesse CI
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: number;
  slug: string;
  name: string;
  description: string;
  icon: string;
  color: string;
  guideCount?: number;
  resourceCount?: number;
}

export interface Guide {
  id: number;
  categoryId: number;
  title: string;
  summary: string;
  stepCount: number;
  duration: string;
  difficulty: string;
  isFeatured?: boolean;
  tags?: string[];
}

export interface GuideStep {
  id: number;
  guideId: number;
  order: number;
  title: string;
  description: string;
  tips?: string[];
  documents?: string[];
  relatedResourceIds?: number[];
}

export type GuideDetail = Guide & {
  steps: GuideStep[];
};

export interface Resource {
  id: number;
  categoryId: number;
  name: string;
  description: string;
  type: string;
  address?: string;
  city: string;
  phone?: string;
  email?: string;
  website?: string;
  services?: string[];
  openingHours?: string;
}

export interface NewsItem {
  id: number;
  categoryId: number;
  title: string;
  summary: string;
  content?: string;
  type: string;
  publishedAt: string;
  deadline?: string;
  source?: string;
  isFeatured?: boolean;
  tags?: string[];
}

export type StatsSummaryGuidesByCategoryItem = {
  categoryId: number;
  categoryName: string;
  count: number;
};

export type StatsSummaryResourcesByTypeItem = {
  type: string;
  count: number;
};

export interface StatsSummary {
  totalGuides: number;
  totalResources: number;
  totalNews: number;
  totalCategories: number;
  guidesByCategory: StatsSummaryGuidesByCategoryItem[];
  resourcesByType: StatsSummaryResourcesByTypeItem[];
}

export type ListGuidesParams = {
  categoryId?: number;
};

export type ListResourcesParams = {
  categoryId?: number;
  type?: string;
};

export type ListNewsParams = {
  categoryId?: number;
  type?: string;
};