# lib/api-zod/src/generated/types/newsItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Jeunesse CI
 * OpenAPI spec version: 0.1.0
 */

export interface NewsItem {
  id: number;
  categoryId: number;
  title: string;
  summary: string;
  content?: string;
  type: string;
  publishedAt: string;
  deadline?: string;
  source?: string;
  isFeatured?: boolean;
  tags?: string[];
}