# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Jeunesse CI
 * OpenAPI spec version: 0.1.0
 */

export interface Resource {
  id: number;
  categoryId: number;
  name: string;
  description: string;
  type: string;
  address?: string;
  city: string;
  phone?: string;
  email?: string;
  website?: string;
  services?: string[];
  openingHours?: string;
}