# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Jeunesse CI
 * OpenAPI spec version: 0.1.0
 */
import type { StatsSummaryGuidesByCategoryItem } from "./statsSummaryGuidesByCategoryItem";
import type { StatsSummaryResourcesByTypeItem } from "./statsSummaryResourcesByTypeItem";

export interface StatsSummary {
  totalGuides: number;
  totalResources: number;
  totalNews: number;
  totalCategories: number;
  guidesByCategory: StatsSummaryGuidesByCategoryItem[];
  resourcesByType: StatsSummaryResourcesByTypeItem[];
}